[IE-132] 사이드바에서 워크스페이스 추가
