--- apps/client/src/components/Layout/components/Header/components/WorkspaceAddButton/hooks/useAddNewWorkspace.ts
@@ -1,11 +1,0 @@
-import useCreateWorkspace from '@hooks/workspace/useCreateWorkspace';
-
-function useAddNewWorkspace() {
-  const createWorkspace = useCreateWorkspace();
-
-  return async (path: string) => {
-    await createWorkspace({ path, name: path });
-  };
-}
-
-export default useAddNewWorkspace;
