+++ apps/client/src/components/Layout/components/Header/components/WorkspaceAddButton/index.tsx
@@ -1,20 +1,11 @@
-import useAddNewWorkspace from '@components/Layout/components/Header/components/WorkspaceAddButton/hooks/useAddNewWorkspace';
-import useInitialPath from '@hooks/file-system/useInitialPath';
-import useOpenFileManagerDialog from '@hooks/useOpenFileManagerDialog';
+import useAddNewWorkspace from '@hooks/workspace/useAddNewWorkspace';
 import AddIcon from '@mui/icons-material/Add';
 import { IconButton, Tooltip } from '@mui/material';
-import { useWorkspace } from 'providers/WorkspaceProvider';
 
 function WorkspaceAddButton() {
-  const openFileManagerDialog = useOpenFileManagerDialog();
-
-  const workspace = useWorkspace();
-
-  const initialPath = useInitialPath();
-
   const addNewWorkspace = useAddNewWorkspace();
 
-  const handleClick = () => openFileManagerDialog(workspace?.path ?? initialPath, addNewWorkspace);
+  const handleClick = () => addNewWorkspace();
 
   return (
     <Tooltip title={'워크스페이스 추가'}>
