+++ apps/client/src/components/Layout/components/Sidebar/components/WorkspaceList/components/WorkspaceAddButton/index.tsx
@@ -1,7 +1,12 @@
 import SidebarMenuAddButton from '@components/Layout/components/Sidebar/components/SidebarMenu/add';
+import useAddNewWorkspace from '@hooks/workspace/useAddNewWorkspace';
 
 function WorkspaceAddButton() {
-  return <SidebarMenuAddButton onClick={() => {}} />;
+  const addNewWorkspace = useAddNewWorkspace();
+
+  const handleClick = () => addNewWorkspace();
+
+  return <SidebarMenuAddButton onClick={handleClick} />;
 }
 
 export default WorkspaceAddButton;
